+++ Sources/KeyboardKit/Keyboard/KeyboardContext+KeyboardType.swift
@@ -33,7 +33,7 @@ private extension KeyboardContext {
         if locale.isRightToLeft { return lowercased }
         switch autoType {
         case .allCharacters: return uppercased
-        case .sentences: return textDocumentProxy.isCursorAtNewSentence ? uppercased : lowercased
+        case .sentences: return textDocumentProxy.isCursorAtNewSentenceWithSpace ? uppercased : lowercased
         case .words: return textDocumentProxy.isCursorAtNewWord ? uppercased : lowercased
         default: return lowercased
         }